+++ packages/mantine/src/components/index.ts
@@ -10,5 +10,6 @@ export * from './table';
 export * from './prompt';
 export * from './modal-wizard';
 export * from './button';
+export * from './action-icon';
 export * from './menu';
 export * from './copyToClipboard';